// sourcecode/BTrak.Packages/snova-hrmangement/module/module-components/src/lib/hrmanagment/models/employee-job-model.ts
export class EmployeeJobDetailsModel {
    employeeJobDetailId: string;
    employeeId: string;
    designationId: string;
    employmentStatusId: string;
    isPermanent: boolean;
    jobCategoryId: string;
    joinedDate: Date;
    departmentId: string;
    branchId: string;
    activeFrom: Date;
    activeTo: Date;
    isArchived: boolean;
    firstName: string;
    surName: string;
    userName: string;
    email: string;
    designationName: string;
    employmentStatus: string;
    jobCategoryName: string;
    departmentName: string;
    branchName: string;
    timeStamp: any;
    totalCount: number;
    noticePeriodInMonths: string;
}